Change products table foreign key constraint datatype

+++ database/migrations/2024_10_22_040523_products.php
@@ -13,13 +13,13 @@ return new class extends Migration
     {
         Schema::create('products', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('seller_id');
+            $table->unsignedBigInteger('seller_id');
             $table->foreign('seller_id')->references('id')->on('users')->onDelete('cascade');
             $table->string('name');
             $table->text('description');
             $table->float('price');
             $table->integer('stocks');
-            $table->unsignedInteger('category_id');
+            $table->unsignedBigInteger('category_id');
             $table->foreign('category_id')->references('id')->on('categories')->onDelete('cascade');
             $table->float('sustainability_score');
             $table->timestamps();
